moved action buttons and related linked from form to stock operation table expandable row
